Move Q(sigma) back into td submodule
