+++ src/agents/control/mod.rs
@@ -13,5 +13,4 @@ pub trait ControlAgent<S: Space, A: Space> {
 
 pub mod td;
 pub mod gtd;
-pub mod mstd;
 pub mod actor_critic;
